refactor: moved header types into header module
