+++ src/rdm/parameter/mod.rs
@@ -4,7 +4,7 @@ pub mod e137_1;
 pub mod e137_2;
 pub mod e137_7;
 
-use super::{RdmError, SubDeviceId};
+use super::{RdmError, header::SubDeviceId};
 
 #[non_exhaustive]
 #[derive(Copy, Clone, Debug, PartialEq)]
